fix bug BILL-6113 + regression test

+++ core/bill_file_handler.py
@@ -1,6 +1,5 @@
 #!/usr/bin/python
 import hashlib
-import requests
 
 from core.model import UtilBill
 from exc import MissingFileError, DuplicateFileError
@@ -84,15 +83,22 @@ class BillFileHandler(object):
             raise MissingFileError('Key "%s" does not exist' % key_name)
 
     def delete_utilbill_pdf_from_s3(self, utilbill):
-        """Removes the pdf file associated with utilbill from s3 (unless
-        there are any other UtilBills referring to the same file).
+        """Remove the file associated with 'utilbill' (unless there are any
+        other UtilBills referring to the same file). If for some reason the file
+        is already missing (e.g. an earlier attempt to delete file caused a
+        transaction rollback after deleting it), nothing happens.
         """
         # TODO: fail if count is not 1?
         if self._utilbill_loader.count_utilbills_with_hash(
                 utilbill.sha256_hexdigest) == 1:
             key_name = BillFileHandler.get_key_name_for_utilbill(utilbill)
             key = self._get_amazon_bucket().get_key(key_name)
-            key.delete()
+            if key is None:
+                # key is already gone
+                # TODO: this error should be logged somewhere...
+                pass
+            else:
+                key.delete()
 
     def upload_file(self, file):
         '''Upload the given file to s3.
